warning eliminated

+++ fs/ide.c
@@ -34,17 +34,17 @@ ide_read(u_int diskno, u_int secno, void *dst, u_int nsecs)
 
 	while (offset_begin + offset < offset_end) {
 		// Your code here
-		syscall_write_dev(&diskno, 0x13000010, sizeof(u_int));
+		syscall_write_dev((u_int)&diskno, 0x13000010, sizeof(u_int));
 		addr = offset+offset_begin;
-		syscall_write_dev(&addr, 0x13000000, sizeof(u_int));
+		syscall_write_dev((u_int)&addr, 0x13000000, sizeof(u_int));
 		a = 0;
-		syscall_write_dev(&a, 0x13000020, sizeof(int));
-		syscall_read_dev(&a, 0x13000030, sizeof(int));
+		syscall_write_dev((u_int)&a, 0x13000020, sizeof(int));
+		syscall_read_dev((u_int)&a, 0x13000030, sizeof(int));
 
 		if(a == 0) {
 			user_panic("ide_read error\n");
 		}
-		syscall_read_dev(dst + offset, 0x13004000, BY2SECT);
+		syscall_read_dev((u_int)dst + offset, 0x13004000, BY2SECT);
 
 		offset += BY2SECT;
 		// error occurred, then panic.
@@ -77,13 +77,13 @@ ide_write(u_int diskno, u_int secno, void *src, u_int nsecs)
 	writef("diskno: %d\n", diskno);
 	while (offset_begin + offset < offset_end) {
 		// copy data from source array to disk buffer.
-		syscall_write_dev(&diskno, 0x13000010, sizeof(u_int));
+		syscall_write_dev((u_int)&diskno, 0x13000010, sizeof(u_int));
 		addr = offset_begin+offset;
-		syscall_write_dev(&addr, 0x13000000, sizeof(u_int));
-		syscall_write_dev(src + offset, 0x13004000, BY2SECT);
+		syscall_write_dev((u_int)&addr, 0x13000000, sizeof(u_int));
+		syscall_write_dev((u_int)src + offset, 0x13004000, BY2SECT);
 		a = 1;
-		syscall_write_dev(&a, 0x13000020, sizeof(int));
-		syscall_read_dev(&a, 0x13000030, sizeof(int));
+		syscall_write_dev((u_int)&a, 0x13000020, sizeof(int));
+		syscall_read_dev((u_int)&a, 0x13000030, sizeof(int));
 		if(a == 0) {
 			user_panic("ide_write error\n");
 		}
